LayoutTests/fast/dom/table-rows-collection-idempotence.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This test verifies that HTMLTableElement.rows returns the same collection when called repeatedly.");

var tableElement = document.createElement("table");

shouldBe("tableElement.rows", "tableElement.rows");
shouldBeTrue("tableElement.rows === tableElement.rows");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
