LayoutTests/fast/dom/table-rows-collection-idempotence-expected.txt

This test verifies that HTMLTableElement.rows returns the same collection when called repeatedly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tableElement.rows is tableElement.rows
PASS tableElement.rows === tableElement.rows is true
PASS successfullyParsed is true

TEST COMPLETE
